requirements.txt
acvl_utils==0.2.5
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
anyio==4.9.0
appnope==0.1.4
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
batchgenerators==0.25.1
batchgeneratorsv2==0.3.0
blosc2==3.3.4
certifi==2025.6.15
charset-normalizer==3.4.2
comm==0.2.3
connected-components-3d==3.23.0
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
Deprecated==1.2.18
dicom2nifti==2.6.1
dynamic_network_architectures==0.4.2
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
fft-conv-pytorch==1.2.0
filelock==3.18.0
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2025.5.1
fury==0.12.0
future==1.0.0
graphviz==0.20.3
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.5
idna==3.10
imagecodecs==2025.3.30
imageio==2.37.0
importlib_resources==6.5.2
ipykernel==6.30.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lazy_loader==0.4
linecache2==1.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==2.0.1
nbformat==5.10.4
ndindex==1.10.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nnunetv2==2.6.2
numexpr==2.11.0
numpy==1.26.4
opencv-python==4.11.0.86
opt-einsum==3.3.0
packaging==25.0
paddlepaddle==3.0.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pydicom==3.0.1
pygltflib==1.16.4
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-gdcm==3.0.25
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
simpleitk==2.5.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.15
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
TotalSegmentator==2.9.0
tqdm==4.67.1
traceback2==1.4.0
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.14.0
tzdata==2025.2
unittest2==1.1.0
urllib3==2.4.0
vtk==9.3.1
wcwidth==0.2.13
wrapt==1.17.2
xvfbwrapper==0.2.13
yacs==0.1.8
yarl==1.20.1